Added WorkingDirectory GetFullDirectoryPath
Added a method for setting a DirectoryInfo WorkingDirectory property so
that a user could set a Working Directory on the SD Card and then have
that Directory Created if it didn't exist.  Also added a static method
to retrieve a full directory path and a method that retrieves the
working directory path.