Implemented WriteLine & Readline
There is an issue in the .Net Micro Framework where it does not truncate
a file when writing contents. I tried opening the file with
FileMode.Truncate and also tried using the objFileStream.SetLength(0)
method. According to the documentation, file truncation is supposed to
occur by default when using FileMode.Create...  The WriteLine and
ReadLine methods will provide a workaround so that valid data can be
read and written before a newline char.